+++ src/steel/types/client_screenshot_params.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from typing_extensions import Literal, Required, Annotated, TypedDict
+from typing_extensions import Required, Annotated, TypedDict
 
 from .._utils import PropertyInfo
 
@@ -19,7 +19,7 @@ class ClientScreenshotParams(TypedDict, total=False):
     full_page: Annotated[bool, PropertyInfo(alias="fullPage")]
     """Capture the full page screenshot. Default is `false`."""
 
-    region: Literal["lax", "ord", "iad", "bom", "scl", "fra", "hkg"]
+    region: str
     """The desired region for the action to be performed in"""
 
     use_proxy: Annotated[bool, PropertyInfo(alias="useProxy")]