+++ tests/api_resources/test_client.py
@@ -33,7 +33,7 @@ class TestClient:
         client_ = client.pdf(
             url="https://example.com",
             delay=0,
-            region="lax",
+            region="region",
             use_proxy=True,
         )
         assert_matches_type(PdfResponse, client_, path=["response"])
@@ -76,7 +76,7 @@ class TestClient:
             delay=0,
             format=["html"],
             pdf=True,
-            region="lax",
+            region="region",
             screenshot=True,
             use_proxy=True,
         )
@@ -119,7 +119,7 @@ class TestClient:
             url="https://example.com",
             delay=0,
             full_page=True,
-            region="lax",
+            region="region",
             use_proxy=True,
         )
         assert_matches_type(ScreenshotResponse, client_, path=["response"])
@@ -166,7 +166,7 @@ class TestAsyncClient:
         client = await async_client.pdf(
             url="https://example.com",
             delay=0,
-            region="lax",
+            region="region",
             use_proxy=True,
         )
         assert_matches_type(PdfResponse, client, path=["response"])
@@ -209,7 +209,7 @@ class TestAsyncClient:
             delay=0,
             format=["html"],
             pdf=True,
-            region="lax",
+            region="region",
             screenshot=True,
             use_proxy=True,
         )
@@ -252,7 +252,7 @@ class TestAsyncClient:
             url="https://example.com",
             delay=0,
             full_page=True,
-            region="lax",
+            region="region",
             use_proxy=True,
         )
         assert_matches_type(ScreenshotResponse, client, path=["response"])